refactor: MemberDataItems 삭제하고 구조 개편
