--- daycan-front/apps/admin/src/pages/member/components/MemberDataItems/index.ts
@@ -1,6 +1,0 @@
-/*
- * MemberDataItems이 컴포넌트는 수급자 목록을 나타내는 컴포넌트 입니다.
- * @author 소보길
- */
-
-export { MemberDataItems } from "./MemberDataItems";
